Move newTextEncoder() and newTextDecoder() to their own file: text_serializer.ts (#7018)
