+++ .changeset/large-lemons-relax.md
@@ -1,0 +1,6 @@
+---
+'@firebase/firestore': patch
+'firebase': patch
+---
+
+Internal refactor of platform-specific logic to create TextEncoder and TextDecoder objects.
