+++ packages/firestore/test/unit/specs/spec_test_runner.ts
@@ -90,7 +90,7 @@ import { Mutation } from '../../../src/model/mutation';
 import { JsonObject } from '../../../src/model/object_value';
 import { encodeBase64 } from '../../../src/platform/base64';
 import { toByteStreamReader } from '../../../src/platform/byte_stream_reader';
-import { newTextEncoder } from '../../../src/platform/serializer';
+import { newTextEncoder } from '../../../src/platform/text_serializer';
 import * as api from '../../../src/protos/firestore_proto_api';
 import { ExistenceFilter } from '../../../src/remote/existence_filter';
 import {
